Fix SYSTEM_POWER_STATUS struct causing access violation
- Correct BatteryLifeTime and BatteryFullLifeTime field types from byte to uint
- This fixes memory corruption in P/Invoke call to GetSystemPowerStatus
- Resolves System.AccessViolationException on application startup

The Windows API specifies these fields as DWORD (uint), not byte.
Using incorrect types caused memory misalignment and crashes.

+++ src/ChargeGuard/Battery/PowerNativeMethods.cs
@@ -17,8 +17,8 @@ public static partial class PowerNativeMethods
         public byte ACLineStatus;
         public byte BatteryFlag;
         public byte BatteryLifePercent;
-        public byte BatteryLifeTime;
-        public byte BatteryFullLifeTime;
+        public uint BatteryLifeTime;
+        public uint BatteryFullLifeTime;
     }
 
     /// <summary>
